Adds IT1_F2_A4 - Building management
Also fixes minor errors in center and university. In addition, this
commit adds control of permissions on lateral menu and on showall
views (displaying or hidding buttons).

+++ Views/AcademicCourse/AcademicCourseShowAllView.php
@@ -1,26 +1,30 @@
 <?php
-class AcademicCourseShowAllView {
+include_once '../Functions/HavePermission.php';
+class AcademicCourseShowAllView
+{
     private $academicCourses;
     private $itemsPerPage;
     private $currentPage;
     private $totalAcademicCourses;
     private $totalPages;
     private $stringToSearch;
-    function __construct($academicCoursesData, $itemsPerPage=NULL, $currentPage=NULL, $totalAcademicCourses=NULL,
-                         $stringToSearch=NULL){
+    function __construct($academicCoursesData, $itemsPerPage = NULL, $currentPage = NULL, $totalAcademicCourses = NULL,
+                         $stringToSearch = NULL)
+    {
         $this->academicCourses = $academicCoursesData;
         $this->itemsPerPage = $itemsPerPage;
         $this->currentPage = $currentPage;
         $this->totalAcademicCourses = $totalAcademicCourses;
-        $this->totalPages = ceil($totalAcademicCourses/$itemsPerPage);
+        $this->totalPages = ceil($totalAcademicCourses / $itemsPerPage);
         $this->stringToSearch = $stringToSearch;
         $this->render();
     }
-    function render(){
+    function render()
+    {
         ?>
         <head>
-            <link rel="stylesheet" href="../CSS/default.css" />
-            <link rel="stylesheet" href="../CSS/table.css" />
+            <link rel="stylesheet" href="../CSS/default.css"/>
+            <link rel="stylesheet" href="../CSS/table.css"/>
         </head>
         <main role="main" class="margin-main ml-sm-auto px-4">
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3">
@@ -35,11 +39,14 @@ class AcademicCourseShowAllView {
                     <a class="btn btn-primary" role="button" href="../Controllers/AcademicCourseController.php">
                         <p data-translate="Volver"></p>
                     </a>
-                <?php else:?>
-                    <a class="btn btn-success" role="button" href="../Controllers/AcademicCourseController.php?action=add">
-                        <span data-feather="plus"></span><p data-translate="Añadir curso académico"></p>
-                    </a>
-                <?php endif;?>
+                <?php else:
+                    if (HavePermission("AcademicCourse", "ADD")): ?>
+                        <a class="btn btn-success" role="button"
+                           href="../Controllers/AcademicCourseController.php?action=add">
+                            <span data-feather="plus"></span>
+                            <p data-translate="Añadir curso académico"></p>
+                        </a>
+                    <?php endif; endif; ?>
             </div>
             <div class="table-responsive">
                 <table class="table table-striped table-sm">
@@ -51,7 +58,7 @@ class AcademicCourseShowAllView {
                         <th><label data-translate="Acciones"></th>
                     </tr>
                     </thead>
-                    <?php if(!empty($this->academicCourses)):?>
+                    <?php if (!empty($this->academicCourses)): ?>
                     <tbody>
                     <?php foreach ($this->academicCourses as $academicCourse): ?>
                         <tr>
@@ -59,12 +66,18 @@ class AcademicCourseShowAllView {
                             <td><?php echo $academicCourse->getStartYear() ?></td>
                             <td><?php echo $academicCourse->getEndYear() ?></td>
                             <td class="row">
-                                <a href="../Controllers/AcademicCourseController.php?action=show&id=<?php echo $academicCourse->getId()?>">
-                                    <span data-feather="eye"></span></a>
-                                <a href="../Controllers/AcademicCourseController.php?action=edit&id=<?php echo $academicCourse->getId()?>">
-                                    <span data-feather="edit"></span></a>
-                                <a href="../Controllers/AcademicCourseController.php?action=delete&id=<?php echo $academicCourse->getId()?>">
-                                    <span data-feather="trash-2"></span></a>
+                                <?php if (HavePermission("AcademicCourse", "SHOWCURRENT")) { ?>
+                                    <a href="../Controllers/AcademicCourseController.php?action=show&id=<?php echo $academicCourse->getId() ?>">
+                                        <span data-feather="eye"></span></a>
+                                <?php }
+                                if (HavePermission("AcademicCourse", "EDIT")) { ?>
+                                    <a href="../Controllers/AcademicCourseController.php?action=edit&id=<?php echo $academicCourse->getId() ?>">
+                                        <span data-feather="edit"></span></a>
+                                <?php }
+                                if (HavePermission("AcademicCourse", "DELETE")) { ?>
+                                    <a href="../Controllers/AcademicCourseController.php?action=delete&id=<?php echo $academicCourse->getId() ?>">
+                                        <span data-feather="trash-2"></span></a>
+                                <?php } ?>
                             </td>
                         </tr>
                     <?php endforeach; ?>
@@ -76,7 +89,7 @@ class AcademicCourseShowAllView {
                 <?php endif; ?>
 
                 <?php new PaginationView($this->itemsPerPage, $this->currentPage, $this->totalAcademicCourses,
-                    "AcademicCourse")?>
+                    "AcademicCourse") ?>
 
             </div>
         </main>
