+++ Views/Functionality/FunctionalityShowAllView.php
@@ -1,25 +1,33 @@
 <?php
-class FunctionalityShowAllView {
+
+include_once '../Functions/HavePermission.php';
+
+class FunctionalityShowAllView
+{
     private $functionalities;
     private $itemsPerPage;
     private $currentPage;
     private $totalFunctionalities;
     private $totalPages;
     private $stringToSearch;
-    function __construct($functionalitiesData, $itemsPerPage=NULL, $currentPage=NULL, $totalFunctionalities=NULL, $toSearch=NULL){
+
+    function __construct($functionalitiesData, $itemsPerPage = NULL, $currentPage = NULL, $totalFunctionalities = NULL, $toSearch = NULL)
+    {
         $this->functionalities = $functionalitiesData;
         $this->itemsPerPage = $itemsPerPage;
         $this->currentPage = $currentPage;
         $this->totalFunctionalities = $totalFunctionalities;
-        $this->totalPages = ceil($totalFunctionalities/$itemsPerPage);
+        $this->totalPages = ceil($totalFunctionalities / $itemsPerPage);
         $this->stringToSearch = $toSearch;
         $this->render();
     }
-    function render(){
+
+    function render()
+    {
         ?>
         <head>
-            <link rel="stylesheet" href="../CSS/default.css" />
-            <link rel="stylesheet" href="../CSS/table.css" />
+            <link rel="stylesheet" href="../CSS/default.css"/>
+            <link rel="stylesheet" href="../CSS/table.css"/>
         </head>
         <main role="main" class="margin-main ml-sm-auto px-4">
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3">
@@ -34,11 +42,14 @@ class FunctionalityShowAllView {
                     <a class="btn btn-primary" role="button" href="../Controllers/FunctionalityController.php">
                         <p data-translate="Volver"></p>
                     </a>
-                <?php else:?>
-                    <a class="btn btn-success" role="button" href="../Controllers/FunctionalityController.php?action=add">
-                        <span data-feather="plus"></span><p data-translate="Añadir funcionalidad"></p>
-                    </a>
-                <?php endif;?>
+                <?php else:
+                    if (HavePermission("UserRole", "ADD")): ?>
+                        <a class="btn btn-success" role="button"
+                           href="../Controllers/FunctionalityController.php?action=add">
+                            <span data-feather="plus"></span>
+                            <p data-translate="Añadir funcionalidad"></p>
+                        </a>
+                    <?php endif; endif; ?>
             </div>
             <div class="table-responsive">
                 <table class="table table-striped table-sm">
@@ -50,7 +61,7 @@ class FunctionalityShowAllView {
                         <th class="actions-row" data-translate="Acciones"></th>
                     </tr>
                     </thead>
-                    <?php if(!empty($this->functionalities)):?>
+                    <?php if (!empty($this->functionalities)): ?>
                     <tbody>
                     <?php foreach ($this->functionalities as $functionality): ?>
                         <tr>
@@ -58,12 +69,18 @@ class FunctionalityShowAllView {
                             <td><?php echo $functionality->getName() ?></td>
                             <td><?php echo $functionality->getDescription() ?></td>
                             <td class="row">
-                                <a href="../Controllers/FunctionalityController.php?action=show&id=<?php echo $functionality->getId()?>">
-                                    <span data-feather="eye"></span></a>
-                                <a href="../Controllers/FunctionalityController.php?action=edit&id=<?php echo $functionality->getId()?>">
-                                    <span data-feather="edit"></span></a>
-                                <a href="../Controllers/FunctionalityController.php?action=delete&id=<?php echo $functionality->getId()?>">
-                                    <span data-feather="trash-2"></span></a>
+                                <?php if (HavePermission("Functionality", "SHOWCURRENT")) { ?>
+                                    <a href="../Controllers/FunctionalityController.php?action=show&id=<?php echo $functionality->getId() ?>">
+                                        <span data-feather="eye"></span></a>
+                                <?php }
+                                if (HavePermission("Functionality", "EDIT")) { ?>
+                                    <a href="../Controllers/FunctionalityController.php?action=edit&id=<?php echo $functionality->getId() ?>">
+                                        <span data-feather="edit"></span></a>
+                                <?php }
+                                if (HavePermission("Functionality", "DELETE")) { ?>
+                                    <a href="../Controllers/FunctionalityController.php?action=delete&id=<?php echo $functionality->getId() ?>">
+                                        <span data-feather="trash-2"></span></a>
+                                <?php } ?>
                             </td>
                         </tr>
                     <?php endforeach; ?>
@@ -71,7 +88,7 @@ class FunctionalityShowAllView {
                 </table>
                 <?php else: ?>
                     </table>
-                    <p data-translate="No se ha obtenido ninguna funcionalidad."> </p>
+                    <p data-translate="No se ha obtenido ninguna funcionalidad."></p>
                 <?php endif; ?>
 
                 <?php new PaginationView($this->itemsPerPage, $this->currentPage, $this->totalFunctionalities,
@@ -88,4 +105,5 @@ class FunctionalityShowAllView {
         <?php
     }
 }
+
 ?>
