+++ Views/Role/RoleShowAllView.php
@@ -1,25 +1,32 @@
 <?php
-class RoleShowAllView {
+include_once '../Functions/HavePermission.php';
+
+class RoleShowAllView
+{
     private $roles;
     private $itemsPerPage;
     private $currentPage;
     private $totalRoles;
     private $totalPages;
     private $stringToSearch;
-    function __construct($rolesData, $itemsPerPage=NULL, $currentPage=NULL, $totalRoles=NULL, $toSearch=NULL){
+
+    function __construct($rolesData, $itemsPerPage = NULL, $currentPage = NULL, $totalRoles = NULL, $toSearch = NULL)
+    {
         $this->roles = $rolesData;
         $this->itemsPerPage = $itemsPerPage;
         $this->currentPage = $currentPage;
         $this->totalRoles = $totalRoles;
-        $this->totalPages = ceil($totalRoles/$itemsPerPage);
+        $this->totalPages = ceil($totalRoles / $itemsPerPage);
         $this->stringToSearch = $toSearch;
         $this->render();
     }
-    function render(){
+
+    function render()
+    {
         ?>
         <head>
-            <link rel="stylesheet" href="../CSS/default.css" />
-            <link rel="stylesheet" href="../CSS/table.css" />
+            <link rel="stylesheet" href="../CSS/default.css"/>
+            <link rel="stylesheet" href="../CSS/table.css"/>
         </head>
         <main role="main" class="margin-main ml-sm-auto px-4">
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3">
@@ -34,11 +41,13 @@ class RoleShowAllView {
                     <a class="btn btn-primary" role="button" href="../Controllers/RoleController.php">
                         <p data-translate="Volver"></p>
                     </a>
-                <?php else:?>
-                    <a class="btn btn-success" role="button" href="../Controllers/RoleController.php?action=add">
-                        <span data-feather="plus"></span><p data-translate="Añadir rol"></p>
-                    </a>
-                <?php endif;?>
+                <?php else:
+                    if (HavePermission("Role", "ADD")): ?>
+                        <a class="btn btn-success" role="button" href="../Controllers/RoleController.php?action=add">
+                            <span data-feather="plus"></span>
+                            <p data-translate="Añadir rol"></p>
+                        </a>
+                    <?php endif; endif; ?>
             </div>
             <div class="table-responsive">
                 <table class="table table-striped table-sm">
@@ -50,7 +59,7 @@ class RoleShowAllView {
                         <th class="actions-row" data-translate="Acciones"></th>
                     </tr>
                     </thead>
-                    <?php if(!empty($this->roles)):?>
+                    <?php if (!empty($this->roles)): ?>
                     <tbody>
                     <?php foreach ($this->roles as $role): ?>
                         <tr>
@@ -58,12 +67,18 @@ class RoleShowAllView {
                             <td><?php echo $role->getName() ?></td>
                             <td><?php echo $role->getDescription() ?></td>
                             <td class="row">
-                                <a href="../Controllers/RoleController.php?action=show&id=<?php echo $role->getId()?>">
-                                    <span data-feather="eye"></span></a>
-                                <a href="../Controllers/RoleController.php?action=edit&id=<?php echo $role->getId()?>">
-                                    <span data-feather="edit"></span></a>
-                                <a href="../Controllers/RoleController.php?action=delete&id=<?php echo $role->getId()?>">
-                                    <span data-feather="trash-2"></span></a>
+                                <?php if (HavePermission("Role", "SHOWCURRENT")) { ?>
+                                    <a href="../Controllers/RoleController.php?action=show&id=<?php echo $role->getId() ?>">
+                                        <span data-feather="eye"></span></a>
+                                <?php }
+                                if (HavePermission("Role", "EDIT")) { ?>
+                                    <a href="../Controllers/RoleController.php?action=edit&id=<?php echo $role->getId() ?>">
+                                        <span data-feather="edit"></span></a>
+                                <?php }
+                                if (HavePermission("Role", "DELETE")) { ?>
+                                    <a href="../Controllers/RoleController.php?action=delete&id=<?php echo $role->getId() ?>">
+                                        <span data-feather="trash-2"></span></a>
+                                <?php } ?>
                             </td>
                         </tr>
                     <?php endforeach; ?>
@@ -75,7 +90,7 @@ class RoleShowAllView {
                 <?php endif; ?>
 
                 <?php new PaginationView($this->itemsPerPage, $this->currentPage, $this->totalRoles,
-                    "Role")?>
+                    "Role") ?>
 
             </div>
         </main>
@@ -88,4 +103,5 @@ class RoleShowAllView {
         <?php
     }
 }
+
 ?>
