+++ Views/UserRole/UserRoleShowAllView.php
@@ -1,25 +1,32 @@
 <?php
-class UserRoleShowAllView {
+include_once '../Functions/HavePermission.php';
+
+class UserRoleShowAllView
+{
     private $userRoles;
     private $itemsPerPage;
     private $currentPage;
     private $totalUserRoles;
     private $totalPages;
     private $stringToSearch;
-    function __construct($userRoleData, $itemsPerPage=NULL, $currentPage=NULL, $totalUserRoles=NULL, $toSearch=NULL){
+
+    function __construct($userRoleData, $itemsPerPage = NULL, $currentPage = NULL, $totalUserRoles = NULL, $toSearch = NULL)
+    {
         $this->userRoles = $userRoleData;
         $this->itemsPerPage = $itemsPerPage;
         $this->currentPage = $currentPage;
         $this->totalUserRoles = $totalUserRoles;
-        $this->totalPages = ceil($totalUserRoles/$itemsPerPage);
+        $this->totalPages = ceil($totalUserRoles / $itemsPerPage);
         $this->stringToSearch = $toSearch;
         $this->render();
     }
-    function render(){
+
+    function render()
+    {
         ?>
         <head>
-            <link rel="stylesheet" href="../CSS/default.css" />
-            <link rel="stylesheet" href="../CSS/table.css" />
+            <link rel="stylesheet" href="../CSS/default.css"/>
+            <link rel="stylesheet" href="../CSS/table.css"/>
         </head>
         <main role="main" class="margin-main ml-sm-auto px-4">
             <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3">
@@ -29,11 +36,14 @@ class UserRoleShowAllView {
                     <a class="btn btn-primary" role="button" href="../Controllers/UserRoleController.php">
                         <p data-translate="Volver"></p>
                     </a>
-                <?php else:?>
-                    <a class="btn btn-success" role="button" href="../Controllers/UserRoleController.php?action=add">
-                        <span data-feather="plus"></span><p data-translate="Asignar rol a usuario"></p>
-                    </a>
-                <?php endif;?>
+                <?php else:
+                    if (HavePermission("UserRole", "ADD")): ?>
+                        <a class="btn btn-success" role="button"
+                           href="../Controllers/UserRoleController.php?action=add">
+                            <span data-feather="plus"></span>
+                            <p data-translate="Asignar rol a usuario"></p>
+                        </a>
+                    <?php endif; endif; ?>
 
             </div>
             <div class="table-responsive">
@@ -45,19 +55,25 @@ class UserRoleShowAllView {
                         <th class="actions-row"><label data-translate="Acciones"></label></th>
                     </tr>
                     </thead>
-                    <?php if(!empty($this->userRoles)):?>
+                    <?php if (!empty($this->userRoles)): ?>
                     <tbody>
                     <?php foreach ($this->userRoles as $userRole): ?>
                         <tr>
-                            <td><?php echo $userRole->getUser()->getLogin() ;?></td>
-                            <td><?php echo $userRole->getRole()->getName() ;?></td>
+                            <td><?php echo $userRole->getUser()->getLogin(); ?></td>
+                            <td><?php echo $userRole->getRole()->getName(); ?></td>
                             <td class="row">
-                                <a href="../Controllers/UserRoleController.php?action=show&id=<?php echo $userRole->getId()?>">
-                                    <span data-feather="eye"></span></a>
-                                <a href="../Controllers/UserRoleController.php?action=edit&id=<?php echo $userRole->getId()?>">
-                                    <span data-feather="edit"></span></a>
-                                <a href="../Controllers/UserRoleController.php?action=delete&id=<?php echo $userRole->getId()?>">
-                                    <span data-feather="trash-2"></span></a>
+                                <?php if (HavePermission("UserRole", "SHOWCURRENT")) { ?>
+                                    <a href="../Controllers/UserRoleController.php?action=show&id=<?php echo $userRole->getId() ?>">
+                                        <span data-feather="eye"></span></a>
+                                <?php }
+                                if (HavePermission("UserRole", "EDIT")) { ?>
+                                    <a href="../Controllers/UserRoleController.php?action=edit&id=<?php echo $userRole->getId() ?>">
+                                        <span data-feather="edit"></span></a>
+                                <?php }
+                                if (HavePermission("UserRole", "DELETE")) { ?>
+                                    <a href="../Controllers/UserRoleController.php?action=delete&id=<?php echo $userRole->getId() ?>">
+                                        <span data-feather="trash-2"></span></a>
+                                <?php } ?>
                             </td>
                         </tr>
                     <?php endforeach; ?>
@@ -69,7 +85,7 @@ class UserRoleShowAllView {
                 <?php endif; ?>
 
                 <?php new PaginationView($this->itemsPerPage, $this->currentPage, $this->totalUserRoles,
-                    "UserRole")?>
+                    "UserRole") ?>
 
             </div>
         </main>
@@ -82,4 +98,3 @@ class UserRoleShowAllView {
         <?php
     }
 }
-?>
